feat: added login, logout, register, and authorization

+++ main/views.py
@@ -1,5 +1,10 @@
+from django.contrib.auth.decorators import login_required  
+from django.core.exceptions import PermissionDenied        
 from django.shortcuts import render
 from django.contrib import messages
+from django.contrib.auth import login, logout
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from django.shortcuts import redirect, render
 from django.core import serializers
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
@@ -7,9 +12,12 @@ from django.shortcuts import get_object_or_404, redirect, render
 from main.models import Experience, Project
 from main.forms import ProjectForm, ExperienceForm
 
+import datetime
+
 
 
 def show_main(request):
+    last_login = request.COOKIES.get('last_login', 'Belum ada sesi login / Cookie tidak ditemukan')
     context = {
         "name": "Faaiz",
         "npm": "2506610286",
@@ -18,6 +26,7 @@ def show_main(request):
             "Mahasiswa Sistem Informasi Universitas Indonesia yang tertarik "
             "pada pengembangan perangkat lunak dan pendidikan."
         ),
+        "last_login": last_login,
     }
     return render(request, "index.html", context)
 
@@ -29,7 +38,11 @@ def show_experience(request):
     }
     return render(request, "experience.html", context)
 
+@login_required(login_url="/login/") 
 def create_experience(request):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+    
     form = ExperienceForm(request.POST or None)
 
     if request.method == "POST" and form.is_valid():
@@ -38,12 +51,17 @@ def create_experience(request):
         return redirect("main:show_experience")
 
     context = {
-        "name": "Burhan",
+        "name": "Faaiz",
         "form": form,
     }
     return render(request, "experience_form.html", context)
 
+
+@login_required(login_url="/login/") 
 def delete_experience(request, experience_id):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+    
     experience = get_object_or_404(Experience, pk=experience_id)
 
     if request.method == "POST":
@@ -65,7 +83,7 @@ def update_experience(request, experience_id):
         return redirect("main:show_experience")
 
     context = {
-        "name": "Burhan",
+        "name": "Faaiz",
         "form": form,
         "experience": experience,
         "is_edit": True,   # handy flag if your template wants to change the title/button
@@ -79,7 +97,7 @@ def get_experiences_json(request):
     if title_query:
         experiences = experiences.filter(title__icontains=title_query)
 
-    experiences_json = serializers.serialize("json", experiences)
+    experiences_json = serializers.serialize("json", experiences, use_natural_foreign_keys=True  )
     return HttpResponse(experiences_json, content_type="application/json")
 
 #keeping the original code just so in case I need to revert back to it
@@ -97,10 +115,14 @@ def get_projects_json(request):
     if title_query:
         projects = projects.filter(title__icontains=title_query)
 
-    projects_json = serializers.serialize("json", projects)
+    projects_json = serializers.serialize("json", projects, use_natural_foreign_keys=True)
     return HttpResponse(projects_json, content_type="application/json")
 
+@login_required(login_url="/login/") 
 def create_project(request):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+
     form = ProjectForm(request.POST or None)
 
     if request.method == "POST" and form.is_valid():
@@ -109,17 +131,72 @@ def create_project(request):
         return redirect("main:show_projects")
 
     context = {
-        "name": "Burhan",
+        "name": "Faaiz",
         "form": form,
     }
     return render(request, "projects_form.html", context)
 
+
+@login_required(login_url="/login/") 
 def delete_project(request, project_id):
+    if not request.user.is_superuser:
+        raise PermissionDenied
+    
     project = get_object_or_404(Project, pk=project_id)
 
     if request.method == "POST":
         project.delete()
         messages.success(request, "Project berhasil dihapus!")
         return redirect("main:show_projects")
+
+    return redirect("main:show_projects")
+
+def register(request):
+    form = UserCreationForm(request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        form.save()
+        messages.success(request, "Akun berhasil dibuat. Silakan login.")
+        return redirect("main:login")
+
+    context = {
+        "name": "Faaiz",
+        "form": form,
+    }
+    return render(request, "register.html", context)
+
+def login_user(request):
+    form = AuthenticationForm(request, data=request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        user = form.get_user()
+        login(request, user)
+        response = redirect("main:show_main")
+        response.set_cookie('last_login', datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
+        return response
+
+    context = {
+        "name": "Burhan",
+        "form": form,
+    }
+    return render(request, "login.html", context)
+
+def logout_user(request):
+    logout(request)
+    response = redirect("main:show_main")
+    response.delete_cookie('last_login')
+    return response
+
+@login_required(login_url="/login/")
+def toggle_star(request, project_id):
+    project = get_object_or_404(Project, pk=project_id)
+
+    if request.method == "POST":
+        # Kalau akun ini sudah pernah memberi star, batalkan star-nya.
+        # Kalau belum, tambahkan star.
+        if request.user in project.starred_by.all():
+            project.starred_by.remove(request.user)
+        else:
+            project.starred_by.add(request.user)
 
     return redirect("main:show_projects")
